// src/types/service.ts
export type service = {
    id: number;
    user_id_get: number | undefined;
    user_id_do: number | undefined;
    type: string;
    title: string;
    description: string;
    category: string;
    skill_level: number;
    hard_level: number;
    created_at: string;
    updated_at: string;
    terminated: boolean;
    picture: string;
};